examples: add a co2mon-tokio example to read the CO2 monitors async
It adds tokio as a dev dependency on Linux (the only async platform right now)
but that shouldn't affect the users of the library.

The existing co2mon example is modified a bit by making a few things public so
we can re-use that code instead of reimplementing this code which is not even
the point of the library.