linux-native: do not panic on invalid strings from the OS
While strings that come from udev should indeed be UTF-8 we cannot expect that
user-provided fields contain anything valid. Make a similar conversion to the C
hidapi version when we encounter an invalid string.

Fields like e.g. the serial are device or user-controlled and there are
UHID libraries out there which fill it in with random values.