Add an async feature skeleton
This gives us the feature and adds the poll methods required to the trait that
backends need to implement.

It also adds a compile error if you try to enable `async` outside of
`linux-native` which will be the only backend to have it initially.

The `async` feature might need to become more internal depending on how the
implementation goes and how far we can stay generic.